requirements.txt
Faker
GitPython
click